[tools] cleanup,minor: shortened shell script.

+++ clean.sh
@@ -2,7 +2,5 @@
 
 BUILD_DIR="_build"
 
-[ -d ${BUILD_DIR} ] || exit 1
-
-rm -rf ${BUILD_DIR}
+[ -d ${BUILD_DIR} ] && rm -rf ${BUILD_DIR}
